requirements.txt
pandas
numpy
minio
scikit-learn
joblib
pyarrow
fastparquet
apache-airflow-providers-cncf-kubernetes
xgboost
mlflow